fix: adding trait and changing small condition
